Tool handler moved from tracing API to observation API
